+++ mediation/AppLovin/source/plugin/Assets/GoogleMobileAds/Api/Mediation/AppLovin/AppLovin.cs
@@ -30,7 +30,12 @@ namespace GoogleMobileAds.Api.Mediation.AppLovin
 
         public static void SetHasUserConsent(bool hasUserConsent)
         {
-            client.SetHasUserConsent (hasUserConsent);
+            client.SetHasUserConsent(hasUserConsent);
+        }
+
+        public static void SetIsAgeRestrictedUser(bool isAgeRestrictedUser)
+        {
+            client.SetIsAgeRestrictedUser(isAgeRestrictedUser);
         }
 
         private static IAppLovinClient GetAppLovinClient()