+++ mediation/AppLovin/source/plugin/Assets/GoogleMobileAds/Common/Mediation/AppLovin/IAppLovinClient.cs
@@ -26,5 +26,9 @@ namespace GoogleMobileAds.Common.Mediation.AppLovin
         // European Union (i.e., EU/GDPR data subject) has provided
         // opt-in consent for the collection and use of personal data.
         void SetHasUserConsent(bool hasUserConsent);
+
+        // Sets a flag indicating whether a user is known to be
+        // in an age-restricted category (i.e., under the age of 16).
+        void SetIsAgeRestrictedUser(bool isAgeRestrictedUser);
     }
 }