+++ mediation/AppLovin/source/plugin/Assets/GoogleMobileAds/Platforms/Android/Mediation/AppLovin/AppLovinClient.cs
@@ -45,11 +45,24 @@ namespace GoogleMobileAds.Android.Mediation.AppLovin
 
         public void SetHasUserConsent(bool hasUserConsent)
         {
-            MonoBehaviour.print("Updating AppLovin user consent flag");
             AndroidJavaClass unityPlayer = new AndroidJavaClass("com.unity3d.player.UnityPlayer");
             AndroidJavaObject currentActivity = unityPlayer.GetStatic<AndroidJavaObject>("currentActivity");
             AndroidJavaClass appLovinPrivacySettings = new AndroidJavaClass("com.applovin.sdk.AppLovinPrivacySettings");
+
+            string parameterString = (hasUserConsent == true ? "true" : "false");
+            MonoBehaviour.print("Calling 'AppLovinPrivacySettings.setHasUserConsent()' with argument: " + parameterString);
             appLovinPrivacySettings.CallStatic("setHasUserConsent", hasUserConsent, currentActivity);
+        }
+
+        public void SetIsAgeRestrictedUser(bool isAgeRestrictedUser)
+        {
+            AndroidJavaClass unityPlayer = new AndroidJavaClass("com.unity3d.player.UnityPlayer");
+            AndroidJavaObject currentActivity = unityPlayer.GetStatic<AndroidJavaObject>("currentActivity");
+            AndroidJavaClass appLovinPrivacySettings = new AndroidJavaClass("com.applovin.sdk.AppLovinPrivacySettings");
+
+            string parameterString = (isAgeRestrictedUser == true ? "true" : "false");
+            MonoBehaviour.print("Calling 'AppLovinPrivacySettings.setIsAgeRestrictedUser()' with argument: " + parameterString);
+            appLovinPrivacySettings.CallStatic("setIsAgeRestrictedUser", isAgeRestrictedUser, currentActivity);
         }
     }
 }