+++ mediation/AppLovin/source/plugin/Assets/Plugins/iOS/GADUMAppLovinInterface.m
@@ -15,9 +15,13 @@
 #import <AppLovinSDK/AppLovinSDK.h>
 
 void GADUMInitializeAppLovin(void) {
-    [ALSdk initializeSdk];
+  [ALSdk initializeSdk];
 }
 
 void GADUMAppLovinSetHasUserConsent(BOOL hasUserConsent) {
-    [ALPrivacySettings setHasUserConsent:hasUserConsent];
+  [ALPrivacySettings setHasUserConsent:hasUserConsent];
+}
+
+void GADUMAppLovinSetIsAgeRestrictedUser(bool isAgeRestrictedUser) {
+  [ALPrivacySettings setIsAgeRestrictedUser:isAgeRestrictedUser];
 }