+++ mediation/AppLovin/CHANGELOG.md
@@ -1,5 +1,10 @@
 # AppLovin Adapter plugin for Google Mobile Ads SDK for Unity 3D Changelog
 
+## 3.0.1
+- Supports AppLovin Android SDK version 8.0.0.
+- Supports AppLovin iOS SDK version 5.0.1.
+- Added `AppLovin.SetIsAgeRestrictedUser()` method to indicate if the user is known to be in an age-restricted category.
+
 ## 3.0.0
 - Supports AppLovin Android SDK version 8.0.0.
 - Supports AppLovin iOS SDK version 5.0.1.